test: fix out-of-range byte literal in SonyResponsesTest [skip ci]
0x99 (153) exceeds the signed Byte range, so byteArrayOf needs an
explicit .toByte(). Test-only; main sources + the release APK are
unaffected.

+++ tool/src/test/kotlin/com/thelightphone/sonywf/protocol/SonyResponsesTest.kt
@@ -95,7 +95,7 @@ class SonyCommandsAndResponsesTest {
 
     @Test
     fun parseReturnsNullForUnknownPayload() {
-        assertNull(SonyResponses.parse(SonyMessage(SonyFrame.TYPE_COMMAND1, 0, byteArrayOf(0x99, 0x00))))
+        assertNull(SonyResponses.parse(SonyMessage(SonyFrame.TYPE_COMMAND1, 0, byteArrayOf(0x99.toByte(), 0x00))))
         assertNull(SonyResponses.parse(SonyMessage(SonyFrame.TYPE_COMMAND1, 0, ByteArray(0))))
     }
 }
